Rewrote room code, added more abstractions.
Broken old channel destruction code (I will rewrite it anyway).

This is a large commit with many changes.
Major additions are Client and Server types.

         Server
         /    \
        /      \
       v        v
  Client <====> Room

Client has references to the channels he subscribed,
Room has references to its subscribers,
Server has references to all Clients and Rooms.

If Client or Room is linked to a server all name changes
should be made using it. If they are not linked then updating
Name property is enough.